refactor: extract photo upload into useProfilePhoto hook
